requirements.txt
inquirer==3.4.0
mariadb==1.1.12
python-dotenv==1.1.0